Move PlainPairing and PlainPlayer to PairingStratgies.
Move build_player_summary to The base pairing strategy and
out of the Tournament model object.
